T-5-1-2 Components refactor, minor code cleanup.
