+++ src/components/layout/nav-menu/nav-menu.style.js
@@ -1,24 +1,74 @@
 // external libraries
-import React from 'react'
+import React, {useContext} from 'react'
 // local services & data store
+import {useRouter} from "../../../services/utilities.service"
+import {NavMenuContext} from '../../../contexts/nav-menu.context'
 // local containers & components
+import Divider from "../../ui/lines/devider"
+import {LineItem, VerticalSheets} from "../../ui/navigation"
+import HeaderArea from "../../ui/navigation/header-area/header-area.component"
+import ToggleSidebar from "../../ui/navigation/toggle-sidebar"
 // local constants & styles
+import CONSTS from "../../../constants"
 import './nav-menu.style.scss'
 
 const NavMenu = () => {
+  const router =  useRouter()
+  const {location: {pathname}} = router
+  const {collapse, setCollapse} = useContext(NavMenuContext)
+
+  const groupAssembler = group => group
+    .map(({LABEL, ICON, ROUTE}) => {
+      const route = Array.isArray(ROUTE) ? ROUTE[0] : ROUTE
+
+      return (
+        <LineItem
+          key={`${LABEL}-${ICON}`}
+          isActive={pathname === route}
+          onClick={() => router.push(route)}
+          label={LABEL}
+          iconLabel={ICON}
+          collapse={collapse}
+        />
+      )
+    })
+
+  const mainGroup = groupAssembler(Object
+    .values(CONSTS.ROUTES)
+    .filter(el => el?.GROUP === 'main'))
+
+  const secondaryGroup = groupAssembler(Object
+    .values(CONSTS.ROUTES)
+    .filter(el => el?.GROUP === 'secondary'))
 
   return (
-    <div className='navMenu'>
-      <div className='navMenu__content'>
-        <span>NavMenu</span>
-        <ul>
-          <li>NavLink</li>
-          <li>NavLink</li>
-          <li>NavLink</li>
-          <li>NavLink</li>
-        </ul>
-      </div>
-    </div>
+    <nav className='navigation-menu horizontal-block'>
+      <VerticalSheets
+        isPlain
+        collapse={collapse}
+      >
+        <HeaderArea
+          title='Sierra Ferguson'
+          isAvatar
+          additionalInfo='s.ferguson@gmail.com'
+          collapse={collapse}
+        />
+        <Divider/>
+        <div className='navigation-menu__group'>
+          {mainGroup}
+        </div>
+        <Divider/>
+        <div className='navigation-menu__group'>
+          {secondaryGroup}
+        </div>
+        <div className='navigation-menu__group'>
+          <ToggleSidebar
+            collapse={collapse}
+            setCollapse={setCollapse}
+          />
+        </div>
+      </VerticalSheets>
+    </nav>
   )
 }
