+++ src/components/layout/nav-menu/nav-menu.style.scss
@@ -1,4 +1,19 @@
+@import "./src/styles/consts";
 
-.navMenu{
+.navigation-menu{
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100%;
 
+  &__group{
+    padding: 12px 8px;
+  }
+
+  & .sub-title{
+    color: $color-black-high-emphasis;
+    &.sub-group{
+      margin-bottom: 10px;
+    }
+  }
 }
