+++ src/components/text-containers/sub-title/sub-title.component.js
@@ -1,6 +1,9 @@
+// external libraries
 import React from 'react'
 import PropTypes from 'prop-types'
-
+// local services & data store
+// local containers & components
+// local constants & styles
 import './sub-title.style.scss'
 
 const SubTitle = ({children, isBold, extraClassname}) => {
