+++ src/components/text-containers/title/title.component.js
@@ -1,4 +1,8 @@
+// external libraries
 import React from 'react'
+// local services & data store
+// local containers & components
+// local constants & styles
 import './title.style.scss'
 
 const Title = ({children, extraClassname}) => {
