+++ src/components/ui/avatar/avatar/avatar.component.js
@@ -1,11 +1,13 @@
+// external libraries
 import React from 'react'
-
+// local services & data store
+// local containers & components
+import Mask from "../mask"
+import Icon from "../../icons/icon.component"
+// local constants & styles
 import womanImg from './Screenshot_20200214_214439.png'
-import Mask from "../../elements/mask"
-import Icon from "../../../icons/icon.component"
 
 const Avatar = ({src = womanImg}) => {
-
   return (
     <Mask>
       {src
