--- src/components/ui/avatar/components/index.js
@@ -1,3 +1,0 @@
-import Avatar from './avatar'
-
-export default Avatar
