--- src/components/ui/avatar/elements/index.js
@@ -1,3 +1,0 @@
-import Mask from './mask'
-
-export default Mask
