+++ src/components/ui/avatar/mask/mask.component.js
@@ -1,15 +1,16 @@
+// external libraries
 import React from 'react'
 import PropTypes from 'prop-types'
-
+// local services & data store
+// local containers & components
+// local constants & styles
 import './mask.style.scss'
 
 const Mask = ({children, width, height}) => {
-
   const style = {
       width,
       height
   }
-
   return (
     <div
       className='mask'
