+++ src/components/ui/avatar/mask/mask.style.scss
@@ -1,4 +1,4 @@
-@import "./src/styles/utilities";
+@import "../../../../styles/utilities";
 
 .mask{
   &__container{
