+++ src/components/ui/lines/devider/divider.component.js
@@ -1,5 +1,8 @@
+// external libraries
 import React from 'react'
-
+// local services & data store
+// local containers & components
+// local constants & styles
 import './divider.style.scss'
 
 const Divider = () => {
