--- src/components/ui/navigation/components/index.js
@@ -1,3 +1,0 @@
-import NavigationMenu from "./navigation-menu"
-
-export default NavigationMenu
