--- src/components/ui/navigation/components/navigation-menu/index.js
@@ -1,7 +1,0 @@
-import NavigationMenu from './navigation.menu.component'
-import NavigationMenuMock from './navigation-menu.mock'
-
-export {
-  NavigationMenu as default,
-  NavigationMenuMock
-}
