--- src/components/ui/navigation/components/navigation-menu/navigation-menu.mock.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import NavigationMenu from "./navigation.menu.component"
-
-const NavigationMenuMock = () => {
-
-  return (
-    <>
-      <NavigationMenu/>
-    </>
-  )
-}
-
-export default NavigationMenuMock
