--- src/components/ui/navigation/components/navigation-menu/navigation.menu.style.scss
@@ -1,19 +1,0 @@
-@import "./src/styles/consts";
-
-.navigation-menu{
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-
-  &__group{
-    padding: 12px 8px;
-  }
-
-  & .sub-title{
-    color: $color-black-high-emphasis;
-    &.sub-group{
-      margin-bottom: 10px;
-    }
-  }
-}
