--- src/components/ui/navigation/elements/header-area/header-area.mock.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import HeaderArea from "./header-area.component"
-
-const HeaderAreaMock = () => {
-
-  return (
-    <>
-      <HeaderArea
-        title='Sierra Ferguson'
-        isAvatar
-        additionalInfo='s.ferguson@gmail.com'
-
-      />
-    </>
-  )
-}
-
-export default HeaderAreaMock
