--- src/components/ui/navigation/elements/header-area/index.js
@@ -1,7 +1,0 @@
-import HeaderArea from './header-area.component'
-import HeaderAreaMock from './header-area.mock'
-
-export {
-  HeaderArea as default,
-  HeaderAreaMock
-}
