--- src/components/ui/navigation/elements/index.js
@@ -1,9 +1,0 @@
-import HeaderArea from './header-area'
-import VerticalSheets from './vertical-sheets'
-import LineItem from './line-item'
-
-export {
-  HeaderArea,
-  VerticalSheets,
-  LineItem
-}
