+++ src/components/ui/navigation/header-area/header-area.style.scss
@@ -1,4 +1,4 @@
-@import "./src/styles/consts";
+@import "../../../../styles/consts";
 
 .header-area{
   padding: 0 16px;
@@ -15,5 +15,8 @@
         color: $color-black-medium-emphasis;
       }
     }
+  }
+  &.collapse{
+    padding: 0;
   }
 }
