+++ src/components/ui/navigation/header-area/index.js
@@ -1,0 +1,3 @@
+import HeaderArea from './header-area.component'
+
+export default HeaderArea
