+++ src/components/ui/navigation/index.js
@@ -1,3 +1,11 @@
-import NavigationMenu from './components'
+import HeaderArea from "./header-area"
+import LineItem from "./line-item"
+import ToggleSidebar from "./toggle-sidebar"
+import VerticalSheets from "./vertical-sheets"
 
-export default NavigationMenu
+export {
+  HeaderArea,
+  LineItem,
+  ToggleSidebar,
+  VerticalSheets
+}
