+++ src/components/ui/navigation/toggle-sidebar/toggle-sidebar.style.scss
@@ -1,4 +1,4 @@
-@import "./src/styles/consts";
+@import "../../../../styles/consts";
 
 .toggle-sidebar{
   width: 100%;
