+++ src/components/ui/navigation/vertical-sheets/vertical-sheets.component.js
@@ -1,13 +1,16 @@
+// external libraries
 import React from 'react'
 import PropTypes from 'prop-types'
 import ClassNames from 'classnames'
-
+// local services & data store
+// local containers & components
+// local constants & styles
 import './vertical-sheets.style.scss'
 
-const VerticalSheets = ({isPlain, children}) => {
-
+const VerticalSheets = ({isPlain, children, collapse}) => {
   const classNames = ClassNames({
-    isPlain
+    isPlain,
+    collapse
   })
 
   return (
@@ -19,10 +22,12 @@ const VerticalSheets = ({isPlain, children}) => {
 
 VerticalSheets.defaultProps = {
   isPlain: false,
+  collapse: false,
 }
 
 VerticalSheets.propTypes = {
-  isPlain: PropTypes.bool
+  isPlain: PropTypes.bool,
+  collapse: PropTypes.bool,
 }
 
 export default VerticalSheets
