+++ src/containers/layout/layout.container.js
@@ -1,13 +1,16 @@
 // external libraries
-import React from 'react'
+import React, {useContext} from 'react'
+import ClassNames from 'classnames'
 // local services & data store
 // local containers & components
 // import Footer from '../../components/layout/footer'
 // import Header from '../../components/layout/header'
 import Main from '../../components/layout/main'
-import NavigationMenu from "../../components/ui/navigation"
+import NavMenu from "../../components/layout/nav-menu"
 // local constants & styles
 import './layout.style.scss'
+import {NavMenuContext} from "../../contexts/nav-menu.context"
+
 
 // header > nav
 // article > section
@@ -15,11 +18,15 @@ import './layout.style.scss'
 // footer
 
 const Layout = ({children}) => {
+  const {collapse} = useContext(NavMenuContext)
+  const className = ClassNames({
+    collapse
+  })
   return (
-    <div className='layout horizontal-block'>
+    <div className={`layout horizontal-block ${className}`}>
       {/* <Header/> */}
       <Main>
-        <NavigationMenu/>
+        <NavMenu/>
         <div className='layout__content'>
           {children}
         </div>
