+++ src/containers/messages/messages.container.js
@@ -1,4 +1,8 @@
+// external libraries
 import React from 'react'
+// local services & data store
+// local containers & components
+// local constants & styles
 
 const Messages = () => {
