+++ src/redux/store.js
@@ -4,10 +4,8 @@ import createSagaMiddleware from 'redux-saga'
 // local services & data store
 import rootReducer from './root.reducer'
 import rootSaga from './root.saga'
-// local containers
-// local components
-// local constants
-// local styles
+// local containers & components
+// local constants & styles
 
 const composeEnhancers =
   typeof window === 'object' &&
