+++ src/styles/consts.scss
@@ -1,4 +1,4 @@
-// fonts
+// fonts6
 $font-main: Roboto;
 $font-fallback: sans-serif;
 $font-weight-thin: 200;
@@ -18,8 +18,7 @@ $width-mobile-md: 479px; // medium-mobile
 $height-block: 40px;
 $width-block: 280px;
 $width-nav-menu-default: 256px;
-$width-nav-menu-collapse: 256px;
-
+$width-nav-menu-collapse: 68px;
 
 // colors
 $color-primary: #2196F3;
